+++ src/controller/model/endpoint.ts
@@ -108,12 +108,11 @@ class Endpoint extends Entity {
                 }
 
                 if (target) {
-                    return {target, cluster: Zcl.Utils.getCluster(entry.cluster)};
+                    return {target, cluster: Zcl.Utils.getCluster(entry.cluster, this.getDevice().manufacturerID)};
                 } else {
                     return undefined;
                 }
-            })
-            .filter((b) => b !== undefined);
+            }).filter(b => b !== undefined);
     }
 
     get configuredReportings(): ConfiguredReporting[] {